+++ native/cocos/base/job-system/job-system-tbb/TBBJobGraph.h
@@ -78,19 +78,19 @@ template <typename Function>
 uint TBBJobGraph::createJob(Function &&func) noexcept {
     _nodes.emplace_back(_graph, func);
     tbb::flow::make_edge(_nodes.front(), _nodes.back());
-    return _nodes.size() - 1u;
+    return static_cast<uint>(_nodes.size() - 1u);
 }
 
 template <typename Function>
 uint TBBJobGraph::createForEachIndexJob(uint begin, uint end, uint step, Function &&func) noexcept {
     _nodes.emplace_back(_graph, [](TBBJobToken t) {});
-    uint        predecessorIdx = _nodes.size() - 1u;
+    uint        predecessorIdx = static_cast<uint>(_nodes.size() - 1u);
     TBBJobNode &predecessor    = _nodes.back();
 
     tbb::flow::make_edge(_nodes.front(), predecessor);
 
     _nodes.emplace_back(_graph, [](TBBJobToken t) {});
-    uint        successorIdx = _nodes.size() - 1u;
+    uint        successorIdx = static_cast<uint>(_nodes.size() - 1u);
     TBBJobNode &successor    = _nodes.back();
 
     for (uint i = begin; i < end; i += step) {
@@ -100,7 +100,7 @@ uint TBBJobGraph::createForEachIndexJob(uint begin, uint end, uint step, Functio
     }
 
     _parallelJobs.push_back({predecessorIdx, successorIdx});
-    return (_parallelJobs.size() - 1u) | PARALLEL_JOB_FLAG;
+    return static_cast<uint>((_parallelJobs.size() - 1u)) | PARALLEL_JOB_FLAG;
 }
 
 } // namespace cc
